refactor(driver): extract x86 window search inputs into a pure tested seam
The four functions that shape how x86 search sees an already-selected
optimization window — the candidate register pool, the enumerative
candidate immediate pool, the per-window live-out contract (with the
conditional-branch narrowing veto), and the non-terminal-Jcc
admissibility gate — lived inline in the ~6.4k-line main.rs binary
crate, interleaved with CLI parsing, ELF I/O, and Capstone bridging.
That is a shallow arrangement: the only way to exercise a rule like
"RSP/RBP are excluded from the pool" or "a trailing terminator vetoes
register narrowing" was to drive a whole search.

Lift them into a new pure library module, `x86_search_inputs`, mirroring
the seam pattern established by candidate_windows (#744) and report
(#743): library types in, library types out; no CLI, ELF, or Capstone.
The run_x86_* drivers become the thin adapter. Each rule is now a
fixture-free unit test in the library rather than a #[cfg(test)] block
trapped in the binary crate.

This also fills a coverage gap: enumerative_immediates_from_target had
no test at all. It now has fixture-free tests (written red-first) pinning
the seed pool [0, 1, -1], first-seen ordering, dedup against seeds and
repeats, and exclusion of register-only instructions.

main.rs: -307 lines. Behavior unchanged; the full ci_check.sh suite
(unit + integration + test_all.sh) passes.
